Update patch set 5
Patch Set 5:

> There isn't /vendor/lib64/libkmcrypto.so on MMB29V, but that if
 > statement is unnecessary because TWRP works fine without it.

Yes, sorry it doesn't exist.
MMB29V is current release of omni bullhead. This patchset is trying to pull in a lib that doesn't exist in MMB29V. How is this supposed to work fine? I'm with jake with the check to avoid "No such file or directory" error when trying to pull in on MMB29V.
Once MHC19J is merged in omni we can safely cleanup the check, we are not gonna list every MHC19J+ vendor lib afterwards.
That being said, it remains my personal humble opinion.

Patch-set: 5
Label: Code-Review=0